requirements.txt
meilisearch
openai
pandas
python-dotenv